[rtree] Factor out benchmarks
