+++ src/bench/mod.rs
@@ -1,1 +1,2 @@
 pub mod btree;
+pub mod rtree;
